fix: make it work for sys services as well
