+++ hydrate-orchestrator/modules/hydrate-kubernetes/render_dump.go
@@ -11,31 +11,6 @@ import (
 	"gopkg.in/yaml.v3"
 )
 
-func (m *HydrateKubernetes) ValidateYamlFiles(
-	ctx context.Context,
-	files []string,
-) error {
-	for _, file := range files {
-		content, err := m.WetRepoDir.File(file).Contents(ctx)
-		if err != nil {
-			return fmt.Errorf("failed to read file %s: %w", file, err)
-		}
-		if err := ValidateYaml([]byte(content)); err != nil {
-			return fmt.Errorf("invalid YAML in file %s: %w", file, err)
-		}
-	}
-	return nil
-}
-
-func ValidateYaml(c []byte) error {
-	var document interface{}
-	err := yaml.Unmarshal(c, &document)
-	if err != nil {
-		return fmt.Errorf("invalid YAML: %w", err)
-	}
-	return nil
-}
-
 func (m *HydrateKubernetes) SplitRenderInFiles(
 
 	ctx context.Context,
@@ -202,6 +177,17 @@ func (m *HydrateKubernetes) DumpSysAppRenderToWetDir(
 
 		}
 
+	}
+
+	renderedFiles, err := m.WetRepoDir.Glob(ctx, "kubernetes-sys-services/"+cluster+"/"+app+"/*.yml")
+	if err != nil {
+		return nil, err
+	}
+
+	err = m.ValidateYamlFiles(ctx, renderedFiles)
+
+	if err != nil {
+		return nil, err
 	}
 
 	return m.WetRepoDir, nil
